GCI 2018: Better layout for creating a new contact
This is part of the Google Code-in 2018 contest.

Implemented changes in the SQLite database adapter and in the `Contact` class. Created and modified methods in order to save all the fields in the contact object correctly, including the photo.
Changes in the Edit/Add contact layout in order to reflect all contact fields. Edited the contact info popup to display all the information (added `ScrollView`) necessary. If there is an empty field, that information is not shown.

Tested using Android Emulator on API 28 (Android 9.0 Pie)